Orders cannot be deleted when start date is equal or later than and end dat is later than today

+++ Car2Go/Car2Go.Services.Data/OrdersService.cs
@@ -42,13 +42,21 @@ namespace Car2Go.Services.Data
         public async Task<bool> Delete(string id)
         {
             var order = ordersRepository.All().FirstOrDefault(x => x.Id == id);
+            if(order.RentStart <= DateTime.Today && DateTime.Today <= order.RentEnd)
+            {
+                throw new InvalidOperationException("Cannot delete a reservation after car already has been rented");
+            }
+            //if (order.RentEnd <= DateTime.Today)
+            //{
+            //    throw new InvalidOperationException("Cannot delete a reservation after car already has been rented");
+            //}
             ordersRepository.Delete(order);
             await ordersRepository.SaveChangesAsync();
             return true;
         }
 
 
-        public IEnumerable<OrderDetailsInputModel> GetOrderById(string id)
+        public IEnumerable<OrderDetailsInputModel> GetOrderById(string userName)
         {
             var orders = this.ordersRepository.AllAsNoTracking()
                 .OrderBy(x => x.User.Email).OrderByDescending(x => x.CreatedOn).
@@ -66,7 +74,8 @@ namespace Car2Go.Services.Data
                     CarGearType = x.Car.GearType,
                     Year = x.Car.Year,
                     Price = x.Price,
-            }).ToList();
+            }).Where(o => o.Email == userName)
+                .ToList();
 
             return orders;
         }
